Boost: Add Cache Preload Queue Management Class
- Implement `Cache_Preload_Queue_Manager` to handle preload queue operations
- Add methods for managing, adding, and processing preload URL batches
- Provide queue management functionality with clear, modular design
